Fixed plotting, added README

+++ src/types.rs
@@ -1,6 +1,6 @@
 use nalgebra::{Const, Dyn, OMatrix, OVector};
 
-pub type FloatType = f64;
+pub type FloatType = f32;
 pub type IndexType = u16;
 pub type RasterType = i16;
